refactor: extract functions to combine deduction and redundancy scripts
